+++ R/map_estimate.R
@@ -1,33 +1,26 @@
 #' Maximum A Posteriori (MAP) Estimate
 #'
-#' Find the \strong{Highest Maximum A Posteriori (MAP)} estimate of a posterior, \emph{i.e.,} the most probable value. It corresponds to the "peak" (or the \emph{mode}) of the posterior distribution. This function returns a dataframe containing the MAP value. If the \code{density} is set to \code{TRUE}, it will include a second column containing the \emph{probability} (\emph{i.e.,} the value of the estimated density function) associated with the MAP (the value of the y axis of the density curve at the MAP).
+#' Find the \strong{Highest Maximum A Posteriori (MAP)} estimate of a posterior, \emph{i.e.,} the most probable value. It corresponds to the "peak" (or the \emph{mode}) of the posterior distribution.
 #'
 #' @inheritParams hdi
 #' @param precision Number of points for density estimation. See the \code{n} parameter in \link[=density]{density}.
-#' @param density Turning this parameter
 #'
-#' @return A numeric value if \code{posterior} is a vector and \code{density = FALSE}.
+#' @return A numeric value if \code{posterior} is a vector.
 #'   If \code{density = TRUE}, or if \code{posterior} is a model-object, returns
 #'   a data frame with following columns:
 #'   \itemize{
 #'     \item \code{Parameter} The model parameter(s), if \code{x} is a model-object. If \code{x} is a vector, this column is missing.
 #'     \item \code{MAP} The MAP estimate for the posterior or each model parameter.
-#'     \item \code{MAP_density}
 #'   }
 #'
 #' @examples
 #' library(bayestestR)
 #'
 #' posterior <- rnorm(10000)
-#'
 #' map_estimate(posterior)
-#' map_estimate(posterior, density = TRUE)
 #'
 #' plot(density(posterior))
-#' abline(v=map_estimate(posterior),
-#'        col="red")  # The x coordinate of MAP
-#' abline(h=map_estimate(posterior, density = TRUE)$MAP_density,
-#'        col="blue")  # The y coordinate of MAP
+#' abline(v=map_estimate(posterior), col="red")
 #'
 #' \dontrun{
 #' library(rstanarm)
@@ -46,62 +39,67 @@ map_estimate <- function(x, ...) {
 }
 
 
+
+
+#' @export
+print.MAP <- function(x, ...) {
+  cat(sprintf("MAP = %.2f", x))
+}
+
+
+
+
+
 #' @rdname map_estimate
 #' @export
-map_estimate.numeric <- function(x, precision = 2^10, density = FALSE, ...) {
+map_estimate.numeric <- function(x, precision = 2^10, ...) {
   d <- stats::density(x, n = precision)
 
   hdp_x <- d$x[which.max(d$y)]
   hdp_y <- max(d$y)
 
-  out <- data.frame(
-    "MAP" = hdp_x,
-    "MAP_density" = hdp_y
-  )
+  out <- hdp_x
+  attr(out, "MAP_density") <- hdp_y
 
-  if (density == TRUE) {
-    out
-  } else {
-    out[names(out) != "MAP_density"]
-  }
+  class(out) <- c(class(out), "MAP")
+  return(out)
 }
+
+
+
+
+
 
 
 #' @importFrom insight get_parameters
 #' @keywords internal
-.map_estimate_models <- function(x, precision, density = FALSE) {
+.map_estimate_models <- function(x, precision) {
   list <- sapply(x, map_estimate, precision = precision, simplify = FALSE)
   out <- flatten_list(list, name = "Parameter")
   rownames(out) <- NULL
 
-  if (density == TRUE) {
-    out
-  } else {
-    out[names(out) != "MAP_density"]
-  }
+  out
 }
 
 #' @rdname map_estimate
 #' @export
-map_estimate.stanreg <- function(x, precision = 2^10, effects = c("fixed", "random", "all"), parameters = NULL, density = FALSE, ...) {
+map_estimate.stanreg <- function(x, precision = 2^10, effects = c("fixed", "random", "all"), parameters = NULL, ...) {
   effects <- match.arg(effects)
 
   .map_estimate_models(
     x = insight::get_parameters(x, effects = effects, parameters = parameters),
-    precision = precision,
-    density = density
+    precision = precision
   )
 }
 
 #' @rdname map_estimate
 #' @export
-map_estimate.brmsfit <- function(x, precision = 2^10, effects = c("fixed", "random", "all"), component = c("conditional", "zi", "zero_inflated", "all"), parameters = NULL, density = FALSE, ...) {
+map_estimate.brmsfit <- function(x, precision = 2^10, effects = c("fixed", "random", "all"), component = c("conditional", "zi", "zero_inflated", "all"), parameters = NULL, ...) {
   effects <- match.arg(effects)
   component <- match.arg(component)
 
   .map_estimate_models(
     x = insight::get_parameters(x, effects = effects, parameters = parameters),
-    precision = precision,
-    density = density
+    precision = precision
   )
 }