+++ NEWS.md
@@ -1,3 +1,12 @@
-# bayestestR 0.1.0
+# Latest
 
-* Added a `NEWS.md` file to track changes to the package.
+### Breaking changes
+
+- `map_estimate` now returns a single value instead of a dataframe and the `density` parameter has been removed. The MAP density value is now accessible via `attributes(map_output)$MAP_density`.
+### New functions / parameters
+### Minor changes
+
+# [0.1.0](https://github.com/easystats/bayestestR/releases/tag/v0.1.0) (2019-04-08)
+
+- CRAN initial publication
+- Added a `NEWS.md` file to track changes to the package.